# perf(es/minifier): Make rename operator parallel