--- a/crates/swc_ecma_transforms_base/src/perf.rs
+++ b/crates/swc_ecma_transforms_base/src/perf.rs
@@ -283,16 +283,18 @@ where
 
 /// This is considered as a private type and it's NOT A PUBLIC API.
 #[cfg(feature = "concurrent")]
-pub trait Items {
-    type Item: Send + Sync;
+pub trait Items:
+    rayon::iter::IntoParallelIterator<Item = Self::Elem> + IntoIterator<Item = Self::Elem>
+{
+    type Elem: Send + Sync;
 
     fn len(&self) -> usize;
 }
 
 /// This is considered as a private type and it's NOT A PUBLIC API.
 #[cfg(not(feature = "concurrent"))]
-pub trait Items {
-    type Item: Send + Sync;
+pub trait Items: IntoIterator<Item = Self::Elem> {
+    type Elem: Send + Sync;
 
     fn len(&self) -> usize;
 }
@@ -306,7 +308,7 @@ pub trait ParallelExt: Parallel {
     fn invoke_par<I, F>(&mut self, threshold: usize, nodes: I, op: F)
     where
         I: Items,
-        F: Send + Sync + Fn(&mut Self, I::Item);
+        F: Send + Sync + Fn(&mut Self, I::Elem);
 }
 
 #[cfg(feature = "concurrent")]
@@ -317,41 +319,55 @@ where
     fn invoke_par<I, F>(&mut self, threshold: usize, nodes: I, op: F)
     where
         I: Items,
-        F: Send + Sync + Fn(&mut Self, I::Item),
+        F: Send + Sync + Fn(&mut Self, I::Elem),
     {
-        {
-            if nodes.len() >= threshold || option_env!("SWC_FORCE_CONCURRENT") == Some("1") {
-                GLOBALS.with(|globals| {
-                    use rayon::prelude::*;
+        if nodes.len() >= threshold || option_env!("SWC_FORCE_CONCURRENT") == Some("1") {
+            GLOBALS.with(|globals| {
+                use rayon::prelude::*;
 
-                    let visitor = nodes
-                        .into_par_iter()
-                        .map(|node| {
-                            GLOBALS.set(globals, || {
-                                let mut visitor = Parallel::create(&*self);
-                                op(&mut visitor, node);
+                let visitor = nodes
+                    .into_par_iter()
+                    .map(|node| {
+                        GLOBALS.set(globals, || {
+                            let mut visitor = Parallel::create(&*self);
+                            op(&mut visitor, node);
 
-                                visitor
-                            })
+                            visitor
                         })
-                        .reduce(
-                            || Parallel::create(&*self),
-                            |mut a, b| {
-                                Parallel::merge(&mut a, b);
+                    })
+                    .reduce(
+                        || Parallel::create(&*self),
+                        |mut a, b| {
+                            Parallel::merge(&mut a, b);
 
-                                a
-                            },
-                        );
+                            a
+                        },
+                    );
 
-                    Parallel::merge(self, visitor);
-                });
+                Parallel::merge(self, visitor);
+            });
 
-                return;
-            }
+            return;
+        }
 
-            for n in nodes {
-                op(self, n);
-            }
+        for n in nodes {
+            op(self, n);
+        }
+    }
+}
+
+#[cfg(not(feature = "concurrent"))]
+impl<T> ParallelExt for T
+where
+    T: Parallel,
+{
+    fn invoke_par<I, F>(&mut self, threshold: usize, nodes: I, op: F)
+    where
+        I: Items,
+        F: Send + Sync + Fn(&mut Self, I::Elem),
+    {
+        for n in nodes {
+            op(self, n);
         }
     }
 }