--- a/crates/swc_ecma_transforms_base/src/perf.rs
+++ b/crates/swc_ecma_transforms_base/src/perf.rs
@@ -283,6 +283,7 @@ where
 
 /// This is considered as a private type and it's NOT A PUBLIC API.
 #[cfg(feature = "concurrent")]
+#[allow(clippy::len_without_is_empty)]
 pub trait Items:
     rayon::iter::IntoParallelIterator<Item = Self::Elem> + IntoIterator<Item = Self::Elem>
 {
@@ -293,6 +294,7 @@ pub trait Items:
 
 /// This is considered as a private type and it's NOT A PUBLIC API.
 #[cfg(not(feature = "concurrent"))]
+#[allow(clippy::len_without_is_empty)]
 pub trait Items: IntoIterator<Item = Self::Elem> {
     type Elem: Send + Sync;
 
@@ -377,6 +379,17 @@ where
     T: Send + Sync,
 {
     type Elem = T;
+
+    fn len(&self) -> usize {
+        Vec::len(self)
+    }
+}
+
+impl<'a, T> Items for &'a mut Vec<T>
+where
+    T: Send + Sync,
+{
+    type Elem = &'a mut T;
 
     fn len(&self) -> usize {
         Vec::len(self)